+++ apps/io-func-sign-events/src/app/config.ts
@@ -1,0 +1,12 @@
+import {
+  ApplicationInsightsConfig,
+  getApplicationInsightsConfigFromEnvironment
+} from "@io-sign/hexagonal-azure-functions";
+
+type Config = {
+  appinsights: ApplicationInsightsConfig;
+};
+
+export const getConfigFromEnvironment = (): Config => ({
+  appinsights: getApplicationInsightsConfigFromEnvironment()
+});