+++ apps/io-func-sign-events/src/application/ports/backoffice-func.ts
@@ -1,0 +1,6 @@
+import type { ServiceUnavailableError } from "@pagopa/hexagonal-core/domain/errors";
+import type { Result } from "neverthrow";
+
+export interface BackofficeFunc {
+  checkHealth(): Promise<Result<void, ServiceUnavailableError>>;
+}