+++ apps/io-sign-backoffice-func/src/infra/http/encoders/webhook.ts
@@ -1,0 +1,19 @@
+import * as E from "io-ts/lib/Encoder";
+
+import { Webhook } from "@/webhook";
+import { WebhookStatusEnum } from "../models/WebhookStatus";
+import {
+  WebhookView as WebhookViewModel,
+  type WebhookView as WebhookViewModelType
+} from "../models/WebhookView";
+
+export const WebhookToApiModel: E.Encoder<WebhookViewModelType, Webhook> = {
+  encode: (webhook) =>
+    WebhookViewModel.encode({
+      url: webhook.url,
+      privateKeySecretName:
+        webhook.privateKeySecretName as WebhookViewModelType["privateKeySecretName"],
+      issuerId: webhook.issuerId as WebhookViewModelType["issuerId"],
+      status: webhook.status as WebhookStatusEnum
+    })
+};