+++ packages/io-core-adapter-azure-functions/src/index.ts
@@ -1,0 +1,6 @@
+export { mountAzureFunctionsRoute, type ErrorResponderConfig } from "./adapters/http.js";
+export {
+  makeAzureTelemetryClient,
+  getApplicationInsightsConfigFromEnvironment,
+  type ApplicationInsightsConfig
+} from "./adapters/appinsights.js";