+++ apps/io-func-sign-events/CHANGELOG.md
@@ -1,1 +1,54 @@
 # io-func-sign-events
+
+## 2.2.0
+
+### Minor Changes
+
+- 9c6a067: Fix esm for azure func
+
+### Patch Changes
+
+- Updated dependencies [9c6a067]
+  - @io-sign/hexagonal-azure-functions@1.2.0
+
+## 2.1.0
+
+### Minor Changes
+
+- bc70889: Moving events-func and hexagonal-azure-functions package to ESM
+
+### Patch Changes
+
+- Updated dependencies [bc70889]
+  - @io-sign/hexagonal-azure-functions@1.1.0
+
+## 2.0.1
+
+### Patch Changes
+
+- 67e5da8: restore HTTP dependency tracking in Application Insights
+- Updated dependencies [67e5da8]
+  - @io-sign/hexagonal-azure-functions@1.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- c1536e7: Implementation of hexagonal pattern
+
+### Patch Changes
+
+- Updated dependencies [c1536e7]
+  - @io-sign/hexagonal-azure-functions@1.0.0
+
+## 1.0.2
+
+### Patch Changes
+
+- c12ce70: Fix env parsing
+
+## 1.0.1
+
+### Patch Changes
+
+- 8bcd156: Configuring setup of application insights